+++ packages/ente-ui/src/redux/selectors.ts
@@ -42,7 +42,7 @@ export const isActionPending: Selector<(action: Action) => boolean> = _.memoize(
  * Auth
  */
 export const isAuthValid: Selector<boolean> = state => {
-  const authState = state.get("auth");
+  const authState = getAuthState(state);
   return authState.cata(
     () => false,
     a => {
@@ -57,10 +57,10 @@ export const isParent: Selector<Maybe<boolean>> = state => {
 };
 
 export const getAuthState: Selector<Maybe<AuthState>> = state =>
-  state.get("auth");
+  Maybe.fromNull(state.get("auth"));
 
 export const getRole: Selector<Maybe<Roles>> = state => {
-  const authState = state.get("auth");
+  const authState = getAuthState(state);
   return authState.map(s => s.get("role"));
 };
 
@@ -73,12 +73,12 @@ export const canCreateEntries = createSelector([getRole], role =>
 );
 
 export const getToken: Selector<Maybe<string>> = state => {
-  const authState = state.get("auth");
+  const authState = getAuthState(state);
   return authState.map(s => s.get("token"));
 };
 
 export const getChildren: Selector<Maybe<UserN[]>> = state => {
-  const authState = state.get("auth");
+  const authState = getAuthState(state);
   return authState.map(s => {
     const childrenIds = s.get("children");
     const children = childrenIds.map(id => getUser(id)(state));
@@ -87,12 +87,12 @@ export const getChildren: Selector<Maybe<UserN[]>> = state => {
 };
 
 export const getDisplayname: Selector<Maybe<string>> = state => {
-  const authState = state.get("auth");
+  const authState = getAuthState(state);
   return authState.map(s => s.get("displayname"));
 };
 
 export const getUsername: Selector<Maybe<string>> = state => {
-  const authState = state.get("auth");
+  const authState = getAuthState(state);
   return authState.map(s => s.get("username"));
 };